Refactor Big numbers and componentize Transaction Details Card
