+++ src/js/Asset.ts
@@ -2,6 +2,7 @@ import api from "@/axios";
 import { IAssetData } from "./IAsset";
 import { profanities } from "@/js/Profanities";
 import Big from "big.js";
+import { stringToBig } from '@/helper';
 
 class Asset {
     id: string;
@@ -11,7 +12,7 @@ class Asset {
     denomination: number;
     name: string;
     symbol: string;
-    volume_day: number;
+    volume_day: Big;
     txCount_day: number;
     profane: boolean;
 
@@ -23,7 +24,7 @@ class Asset {
         this.denomination = assetData.denomination;
         this.name = assetData.name;
         this.symbol = assetData.symbol;
-        this.volume_day = 0;
+        this.volume_day = Big(0);
         this.txCount_day = 0;
         this.profane = false;
         this.updateVolumeHistory();
@@ -41,7 +42,7 @@ class Asset {
         api.get(`/x/transactions/aggregates?startTime=${startDate.toISOString()}&endTime=${endDate.toISOString()}&assetID=${this.id}`).then(res => {
             let txCount = res.data.aggregates.transactionCount;
             let txVolume = res.data.aggregates.transactionVolume;
-            parent.volume_day = parseInt(txVolume) / Math.pow(10, parent.denomination);
+            parent.volume_day = stringToBig(txVolume, parent.denomination);
             parent.txCount_day = txCount;
         });
     }
